service_desk_backend/requirements.txt
Flask
PyMySQL
PyYAML